+++ shared/src/canvas/index.ts
@@ -4,4 +4,4 @@
 
 export { canvasReducer, initialState, LIVE_MESSAGE_ID, MAX_MESSAGES } from './reducer';
 
-export type { CanvasAction, CanvasHookState } from './reducer';
+export type { CanvasAction, CanvasHookState, PendingStrokesInfo } from './reducer';